Added check before deleting a file to the controller

+++ frontend/controllers/MessageController.php
@@ -150,8 +150,11 @@ class MessageController extends Controller
             {
                 foreach ($fileNames as $fileName)
                 {
-                    $file = Yii::getAlias('D:/Develop/eshop/frontend/uploads/'."$fileName");
-                    unlink($file);
+                    if(file_exists('D:/Develop/eshop/frontend/uploads/'."$fileName"))
+                    {
+                        $file = Yii::getAlias('D:/Develop/eshop/frontend/uploads/'."$fileName");
+                        unlink($file);
+                    }
                 }
                 return;
             }
